Fixes place, placeId, place_id bug. Separates graphql definitions and resolvers.
