+++ nodemon.json
@@ -2,7 +2,7 @@
   "watch": [
     "src"
   ],
-  "ext": "ts",
+  "ext": "ts,graphql",
   "ignore": [
     "src/**/*.test.ts"
   ],
