+++ src/gql/resolvers/__tests__/CommonResolvers.test.ts
@@ -1,6 +1,6 @@
-import { schema } from '..'
+import { schema } from '../../'
 import { graphql } from 'graphql'
-import { createRepositoryMock, createProfile } from '../../data/entities/__mocks__/'
+import { createRepositoryMock, createProfile } from '../../../data/entities/__mocks__/'
 
 describe('Place - GraphQL Definitions and Resolvers', () => {
   const fake: any = {}
