+++ src/gql/resolvers/__tests__/PlaceResolvers.test.ts
@@ -1,6 +1,6 @@
-import { schema } from '..'
+import { schema } from '../../'
 import { graphql } from 'graphql'
-import { createRepositoryMock, createPlace } from '../../data/entities/__mocks__/'
+import { createRepositoryMock, createPlace } from '../../../data/entities/__mocks__/'
 
 describe('Place - GraphQL Definitions and Resolvers', () => {
   const fake: any = {}
