+++ src/gql/types/common.graphql
@@ -1,0 +1,23 @@
+scalar Date
+
+input RelationId {
+  id: Int!
+}
+
+type DBResponseRaw {
+  fieldCount: Int
+  affectedRows: Int
+  insert: String
+  serverStatus: Int
+  warningStatus: Int
+}
+type DBResponse {
+  raw: DBResponseRaw
+}
+
+type Query {
+  _empty: String
+}
+type Mutation {
+  _empty: String
+}
