+++ src/gql/types/place.graphql
@@ -1,0 +1,26 @@
+enum PlaceType {
+  country
+  state
+  region
+  city
+  district
+}
+input PlaceInput {
+  name: String
+  type: PlaceType
+}
+type Place {
+  id: Int
+  name: String
+  type: PlaceType
+}
+
+extend type Query {
+  place(id: Int!): Place
+  places(first: Int, offset: Int): [Place]
+}
+extend type Mutation {
+  createPlace(input: PlaceInput): Place
+  updatePlace(id: Int!, input: PlaceInput): Place
+  deletePlace(id: Int!): DBResponse
+}
