+++ src/gql/types/profile.graphql
@@ -1,0 +1,29 @@
+input ProfileInput {
+  email: String
+  firstName: String
+  lastName: String
+  birthDate: Date
+  address: String
+  phone: String
+  place: RelationId
+}
+type Profile {
+  id: Int
+  email: String
+  firstName: String
+  lastName: String
+  birthDate: Date
+  address: String
+  phone: String
+  place: Place
+}
+
+extend type Query {
+  profile(id: Int!): Profile
+  profiles(first: Int, offset: Int, after: Int): [Profile]
+}
+extend type Mutation {
+  createProfile(input: ProfileInput): Profile
+  updateProfile(id: Int!, input: ProfileInput): Profile
+  deleteProfile(id: Int!): DBResponse
+}
